fix: Use custom reporter to flip expected failures
To verify that the new `fail` matcher works correctly, we need a way to
check if the tests failed as expected or not.

Typically one would use `test.failing` for this purpose, but that only
works if the test threw an exception (e.g. JestException). The `fail`
matcher does not do this so that it can still work inside `catch`
blocks.

Instead, we have to hook into the test reporting and mutate the test
results for our expected test failures prior to usage by other test
reporters.

This commit creates a custom test reporter which detects tests run in
the file `fail.test.js` (yes the name is hard-coded) and flips the
results of tests inside a `.fail` describe block (i.e. our expected
failures).
 - If the tests failed (expected) we flip the result to a pass.
 - If the tests passed (unexpected) we flip the result to a fail.

The custom reporter also handles the logic for updating the counts for
failing test suites so that later reporters reflect the actual test
results correctly.